Unregister on Destroy
The SC3Editor wasn't unregistering itself as a message target
or log target when destroyed, with deleterious effects.
Reaper mac wouldn't work on a second open for instance.
I also suspect this will cure the savi on startup crash
in #79

While cleaning up I also added a non-virtual sendAD
function so we can avoid calling a virtual on self
in our constructor and destructor, and make the interface
function a passthrough.