Beginning a path to engine cleanup
The engine was ported and working, but the code had never really
been visited, analyzed, cleaned up, etc... and that's kinda what
I want to do this week. Start by deleting some distracting stuff
which was unused and cleaning up some of the expedient hacks
from the very first port.