Day 45: Sample Loops abut the End (#514)
If a sample loop had its last point be the end wavepoint
the FIR reached into the 0 padding at the loop leading to a
click. Fix this by adding an indirection pointer which almost
all the time is a direct pointer into the sample but at that loop
endpoint, fills with the wrapped-sample-set for the FIR.

Tnis will let us do loop fades similarly in the future
(namely make a fake input buffer which is the fade of the samples
since the combination is linear) so let me tag #439 here also.