Day 46: Prep for MultiSelect; Add voice state to UI (#515)
This is basically a structural-only commit which does a couple of
larger things

First it adds a structure mutation mutex. Since only the running
engine or the serialization thread with a paused engine can mutate
the state, this is *not* locked in audio playback but is locked when we
add zones and so on. But it means that the serialization thread can
lock it so it can always get a coherent view of the structure
(but maybe not the values in the sturcture). Good.

Then add a voice status structure which we use to communicate
which zones are playing and which midi notes are pressed to the
UI but for now just show those in the log don't actually repaint
based on them.