Move from ControlDescription to bbParamMetaData
sst::basic_blocks::params::ParamMetaData is where we are headed
so the internal ControlDescription we had in short circuit as a
place holder bites the dust and we use that class (which is
usinged as datamodel::pmd here) everywhere consistently.

Lots more to do, but this works and couples appropriately.