Add a shared memory read only direct message region (#544)
For some non-structural things, like vu meters, sample
position, and playing voices, we don't really need to
send json messages, but instead can just make a shared
memory region with atomics which is const & loaded
in the editor. But placing this in one structure we
know what we would lose in a fully remote situation
(like a far away web browser). But we can also send
higher frequency updates with asimple tail chasing
int64_t. Closes #535

Along the way, put the voice info into the
structure, which stops that little update chase we
had. But a different way. Closes #531