Deadlock absent Audio on File Drop
If a file was dropped and we had no audio engine so we ran
the structure mutation on the serialization thread, we recursively
locked the non-recursive structure mutex. rather than face the
horrors of recursive mutexes, add a new callback to separate
messages based on locking or not on the engine, and have the
no engine case use the existing lock for both.

Closes #547