Utility Merge: Browser API plumbed in
This is a merge with no new user functionality other than a deceptive and not very useful display, but it adds a class for the browser, sets up ownership, pushes a const& to the UI, starts some API, sets up platform dependent bits and bobs. That sort of stuff. Basically "pre-plumbing".