Use the SQLIte for Favorite Locations (#605)
If you have a location in the browser, you can add it as a
location which shows up in your devices. This is of course
no where near a final design but it totally round trips the
sql -> threading -> ui -> filesystem model and also
shows how to set up all the messages and so forth.