A bunch more work towards output editing and routing (#634)
Didn't quite get the knob on the screen but managed to
get group and zone output info up to the client and the
proper object model for both. Merge it just to keep
current. Put the controls on the screen tomorrow (and
maybe add a proc routing dropdown while i'm there)

Also add proc routing enum to the output infos
but don't use it for anything yet. Will get that tomorrow
or saturday