Fix Logic Pro Startup / MIDI VM error
The voice manager assumes well formed midi messages of size
three, but that means if you send it a sysex blast, you can get
a blammo. Logic does that on startup with a mesage I didn't
debug, but basically this change makes it so only 3 byte
midi 1 messages get through the wrapper for now.

That stops logic from crashing on startup

Closeds #656