Checkpoint towards Fancy UIs
This contains some more work towards fancy uis but they are off
in this commit. I'm committing it since I found a DSP error in the
eq port from yesterday and want to get that in in case people download
over holiday break and I don't get another commit.

Main changes here are

- sst effects update with fixes
- processor pane even more branching for per-processor uis
- an eq n band ui which runs the eq frequency graph