OneShot moves to correct storage location
Curves don't have a 'oneshot' - it is a feature of steps
and msegs. Also it is distinct from trigger mode. You could
ahve a release oneshot for instance. So move it back to
the Step storage and move the button from the mutliswitch

Oh also make sure it works. It does!