Rework the mod matrix
The mod matrix was by far the worst designed part of the codebase.
This commit fixes it by introcucing new classes and identifiers
moving the core implementation to basic blocks and generally
being more rational and efficient in where all the code is,
including display names.

Still a lot to do - notably this commit breaks multiselect
mod matrices - but a great checkpoint.