Handle state load threading issue
State load had always been a bit thread dicey. it turned off
the thread checker for goodness sake! So lets do the right
thing and actually use the message queue to load state, stopping
audio, etc...

With this I can load a group processor project 10 times in reaper
VST3 or in the clap or juce standalones and get it working no
problem. So I think this addresses #824