Modify Zone Sample Data; Introduce Round Robin modes
the zone sample data was just an array of samples leaving
no place to store collections of data edited by the sample view.
So change it to a struct and add a variation playback mode enum
to that struct.

Then support that enum and push it to the ui (with a new
single field update sample message which we should use for
other things like loops and stuff - I'll open an issue)

And then implement the four round robin modes.