Processor Invalidation; Keytrack; SuperSVF Gone (#858)
With features like sst-filters (t/k) or keytrack (this commit)
a change to an int or a bool param can change the nature and
value of parameters. This commit begins to support that by handling
the keytrack case. It also adds Keytrack to the Fast SVF.

In order to do that, we need to remove the old super svf, which is
redundant with sst-filters, and doing so also allows a lot of cleanup
on pre-voice-effects processors, moving now all our processors to the
voice effects shim style.

While there's still some message-size-efficiency work to do and
while we still need a processor-int-can-change-values-not-just-types
hook, this is a really solid check point, so commit here.