Change the meaning of 'fadeStart/fadeEnd'; adjust (#884)
Previous versions had fadeStart and fadeEnd be the midi note
of the fade regions. This led to lots of complexity lots of
spaces, including not making '0' a good default.

So change the internal meaning so fadeStart and fadeEnd are the
distance from start and end of the fade zone. Once you do this
lots of stuff gets simpler

- ui constraints are easier to calculate
- initialization is fine at 0
- you can move the fade calculation onto the range objects easily
- etc...

While in there also add an 'addGuiStepPreSend' to the connectors
so you can do constrained guis more easily.