Make processor mix defaultable per-type
Basically add a default mix customization based on type;
use that to generate the metadata (so the metadat agenerator
can inspect the payload which is useful anyway) and template
type vary it. Then add a base class and specialization macro.
Then set a couple to make sure it works and turn it over to
andreya to set them all.

Closes #888