Proc ROuting Step 1; Also Clang-Format (#939)
Proc Routing step 1: Move the routing to a shared function for
linear and share it between group and voice.

Also I ran my clang formatter and it found lots of bad formatting
and missing headers, so that's all swept in also. I wish it was
separate but I can't unpick it now I did it. Sigh.