Generator wasn't accounting for FIR padding
When samples load we load them with an 8-sample buffer on
either side then use GetSamplePtrBlah to get the top of the
actual sample buffer, so we can read "out of bounds" safely
(namely sample[-1] is allocated memory set to zero and is not
actually out of bounds).

But for that to work, you need to tell the generator to read
samples from the top of the waveform not the zero point

This caused clicns and fades in perfect loop samples. Remediate.

Also clean up some distracting logs.

Closes #953