Improve the fade calculation a smidge (#987)
The fade calculation cubed, as opposed to cube rooted,
gain->amp and also did it before subtraction defeating the
purpose. So change the fade calculation to have a standalone
fade gain to amp calculation (which here we use an approx to
cbrt which works pretty well) and then apply it after the
f, 1-f not before.

This makes fade loops sound way less uneven. Still some focus
and debugging I'm sure but generally better.