Towards temposync step 1 (#989)
I am pushing towards temposync working everywhere but it requires
some tweaks so this is just a small incremental step that allows
you to edit the temposync state on an LFO, even though we dont
do anything with that edited state yet. checkpointing into main though
so I can now do the real surgery around it.