Save Multi, Improve User Paths, and More (#1070)
This commit is the first commit which allows you to save
a multi (without embedded samples at this time) to a file
and then load it later cleanly in a running session. It
contains a wide set of changes.

The basic design though is to save a RIFF file with state
and associated data. This is all an internal format in patch_io
writing RIFF with the libgig RIFF tools.

As well as those changes, to get it working at the same time I
made the following:

- User data paths are now on the browser
- Browser can determine loadable single vs multi vs sc files
- Update streaming version and fix enum streaming
- Change the drag and drop logic
- Add multiple isLoadable methods to the browser
- Add drop on header of shortcircuit file to load
- Tuning and Zoom Menu in place
- All other buttons pop 'coming soon'
- Clear engine function implemented