A clap first juce based sc only standalone
While waiting and working on clap wrapper standalone put
in place a temporary hacked version of the juce standalone
filter window specific to scxt which doesn't require an audio
processor intermediate nor juceaide.

This commit turns it on but doesn't make it a default. Stil work
to do on installer and testing on linux and windows (which partly
this PR does).