Add client notification messages
Per #1159 we now at least have a serialization thread message
which the engine can send to the UI with an open/close/update
type semantic. We send the messages from SF2 load but we don't do
anything with them other than a SCLOG_ONCE for now. Just some
progress to get in the codebase before I close my laptop.