An experiment with porting an importer from ConvertWithMOSS
This is an experiment which we may not finish for 1.0 with
porting an importer from ConvertWithMoss, in this case the
EXS importer. It kinda shows the approach but it doesn't
work very well right now since

1. I don't pick apart the structure carefully and
2. Logic has expanded the format ahead of CWM some and
3. It's just a lot of fiddly work

So I'll commit this and perhaps nuke it from the code base
later ,but I didn't want to abandon the couple of hours of work
so merge it for now.