Group Processor Ringout support
This adds support for a processor to implement tailLength
and then keep a group alie through its tail. it also adds a
reasonable tailLength implementation to ShortDelay (but no
other processor).

Addresses #852. Once we do the same with EG12 release that
issue will be mostly wrapped, but we do need to visit all the procs for
a tail.