Towards better sample zoom
This is a more scalable way to do sample zoom. Instead of a
juce viewport with a growing component, we have an api to
collaborate with the component and it chooses what to draw into
a scaled window.

This also means I've consistently gone to either unit-axis or
sample-axis coordinates and have conversions back and forth and
so on.

Three gotchas still to do

1. I tested all the gestures with a trackpad and they are great but do need to test them with a regular mouse wheel. I"m pretty sure they will be wonky there
2. I haven't turned on vertical zoom just horizontal zoom.
3. I haven't improved the painting to look like wireframe