Temporary workaround for accidental move on select
In  bd2b2e4 we made select start a drag event but this meant
tiny mosue movements when selecting would send a drag event simultaneously
with the select event on the wrong zone, making zones overlap.

The correct fix for this is to wait for the new selected zone or update
the view internally when selecting, so this is not a good solution
long term, buit this fixes the bug in the interim.

Addresses #1270