Lift bus effects to voice (#1293)
this commit allows bus-to-voice effects to be used
on group processors. It also introduces an api for
group-only processors. It brings in R1, R2, Delay and
Flanger. It updates the memory pool API

It does not support temposync, and it usese the
underlying tail apis, which for delay at least is
incorrect. I'll open a new issue for those. But this
can close a couple of others

Closes #1280
Closes #1281