Starting on depth slider throw scaling
This commit injects a rescaling data adapter between the depth
slider and the underlying depth which basically uses a cube / cuberoot
on the depth slider. This is part of the plan for #767 but not all of it

With this a 1 semitone pitch adjustment is about 25% of the total throw
so it is already a vast improvement, so merge it for now to make the
synth generaly way less annoying to use.