Mod Matrix Smoothing in place (still defaults to off)
This commit implements smoothing, both linear and exponential,
in the mod matrix. It exposes this in the UI as a sub-menu for
a source and it is available for every source (except of course
for "No Source").

Currently there's no per-source defaults for the matrix.
I'll add that as a separate issue but lets use this commit to
close #1343