Towards basic missing sample resolution
Missing samlpes get placed in the sample hierarchy as types
we can reason about and replace. This allows us to detect which
zones have missing samples and beam that to the UI so we can color
things at display time as well as just showing the missing list
error dialog.

Along the way clean up some pragmatic use of tuples and replace
them with some actual structures that stream.

This still doesn't let us *fix* anything. But it lets us identify
the problem spots at least in the UI. More to come soon obviously