Restructure SF2 to not over-load
The SF2 indexes samples by the preset/instrument/region
but that resolves to a sample index. Since I was using P/I/R
as SampleID that meant identical samples had different IDs
and thus loaded multiple times. Defacto this made a 3MB SF2
with lots of instruments load as 90MB and stuff. Ugh.

So fix this in two ways

1. Make loadeFromSF2 (p,i,r) remap the sample to index if needed
2. Change the sample->load to be index based

and voila