Voice Stealing and Termination Improvements (#1414)
* Voice Stealing and Termination Improvements

Handle a voice stealing API which, for now, steals only at
the hard limit, but has been tested at both hard and soft
limit at this commit with a different configuration. Closes

At the same time, add a voice termination fade for soft
terminates which keeps the voice alive for 8 blocks with
a fade during the period. Closes #801

* Closes #1392