Fix a problem with streamed file names Win vs Posix (#1516)
Streamed file names were in the native format. This meant
that a windows streamed SCM didn't even get to the resolver
for missing samples.

In the future, stream in preferred form, and that will help
(but not obviate this commit). For now fix on unstream which
is sort of a minimum fix.

of course relative paths will change this code a lot, hence the
pause on the preferred stream