Commit for pull request for interpolation
Implemented a new interpolation mode for Shortcircuit XT. A variation of the custom interpolation used in my sfizz fork: sonically, it sounds pretty passable for an early variable-clock DAC sampler: it's like piecewise interpolation but without the excessive aliasing, hence, the title I gave it in the code, but functionally, this additional mode is a modified version of the included sinc interpolation. My changes are rough around the edges, most likely, so revision suggestions from others are welcome.