Redo of commit and pull request
Redo of pull request #1625 in light of a bad codebase

A bit of a code rewrite for the new interpolation, admittedly, in light of concerns from others, which I myself can see validity to insofar as performance.

Instead of using sinc, basis interpolation was "downgraded" to a new cubic interpolation, though the crucial subpos math still applies, albeit now without the "max" function in light of some wanting aliasing for the high notes. This also means this PR is open to revisions regarding what to name the new type of interpolation, which I kept as is as a placeholder.