Allow processors to data-advertise enabled params in matrix (#1634)
This commit allows

1. ParamMetadata to have an isEnabled state
2. If that is false, the matrix target list for that parameter
   is not enabled / grays out (rather than vanishes)
3. Updates these fields properly on an int change if you need
   a consistency check
4. Adds a consistency check to the VA oscillator

This means, in short, that the non-used fields like pulse-width-in
sine-mode on teh VA oscillator show up as non-enabled fields in the
mod matrix (and if you have one from pulse times and switch to sine
the label grays out in the row). But the general feature of doing
this for params outside the VA is also in place.

Closes #1027