Baseline on how far I got with exs; add exs-dump cli (#1648)
In pursuit of #1647 I spent a few minutes this morning
remembering how far I got with exs support. Answer was actually
pretty far, just it doesn't import yet. Also no param parsing.

But to test that I added a little exs-dump command so might
as well keep that in the codebase. So just a placeholder commit
as I scan issues and make sure we aren't missing things