"MOve Selected Groups To New Group" RMB
As part of restructuring these move and sidebar messages
I can just pop in a quicl proof of concept for move to new group
in the RMB on both sidebar and mapping pane.

Closes #1201
Addresses #1653 (which really subsumes #1201)

This is just a tiny subset of the RMB but it lets me
close off that issue and ship the feature at least.