Unused Busses optimization, Hotel California edition (#1714)
This adds an optimization where unused busses don't process.
An unused bus is one in this sense which has *never* been
written to since the patch reset. Hence the name
"Hotel California". You can check into the audio graph
any time you want, but you can never leave.

So what we need to do, and what I'll change #1223 to be, is
to instead have a tail calculation based on the active effects
and whenever written (that is when we do busUsed to true) we
set busUsed to the tail length and then count down.

So this only addresses #1223, but it does the plumbing so
we can do the rest.