Fix a problem in group tail implementation (#1902)
The group tail implementation stopped the group correctly
but didn't notify the parent that the group was stopped.
Fix that through the implementation of an if rather than
else in the group chain.

I figured this out by starting to work on part ringout so
did a bit of the pre-work for that but nothing material.