Step 1 of Bus Ringout (#1904)
This is step 1 of bus ringout. The bus now knows if it is
in a potential ringout state (namely nothing has been pushed
to it) so it can self-shortcircuit. It doesn't do that though
and just keeps processing, here with a turned-off log for
when it is in ringout state

Addresses #1223