A step along the way to the phasor gui (#1909)
This is a step along the way to almost being done with #1643

It introduces a draggable gui for note/dot/triplet mode which
goes in the 'valhallapilled' style. It allows typeins.

There's two big problems

1. It no longer matches the wireframe. Custom paint needed. Fine.
2. When you switch into and out of OF_BLAH mode you can set a fraction
   which when you resturn to triplet mode shows as the nearest on
   screen without engine updated

Both of those are needed to close the issue. But I want to merge
this so Andreya can try it in the AM