Lock the samples map when accessed or mutated
The UI thread would talk to the sample database. In the event the
sample database was being purged or reaped that meant the UI could
just be talking to an incomplete sample std::unordered_map

Fix it by having a lock-for-map-traversal idiom.

The contention here for a zone accessing the sample is all
on the serialization thread so should be audio safe and
cricitcally this lock isn't required to start a voice.

Closes #2043
Addresses #2034